+++ src/offers/offers.module.ts
@@ -2,10 +2,12 @@ import { Module } from '@nestjs/common';
 
 import { PrismaModule } from '../prisma/prisma.module';
 import { OfferOwnerGuard } from './guards/offer-owner.guard';
+import { OffersController } from './offers.controller';
 import { OffersService } from './offers.service';
 
 @Module({
   imports: [PrismaModule],
+  controllers: [OffersController],
   providers: [OffersService, OfferOwnerGuard],
   exports: [OffersService, OfferOwnerGuard],
 })